backend/requirements.txt
# Python dependencies
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# Twitter Scraping
snscrape==0.7.0.20230622

# Database
pymongo==4.5.0
dnspython==2.4.2

# Graph Analysis
networkx==3.1
python-louvain==0.16
scikit-learn==1.3.0

# Data Processing
pandas==2.1.1
numpy==1.25.2

# Scheduling
APScheduler==3.10.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0